+++ benches/algo/crosscountry.rs
@@ -1,4 +1,4 @@
-//! Benchmark [`graaf::algo::dijkstra::distances`] for different digraph
+//! Benchmark [`graaf::algo::dijkstra::min_distances`] for different digraph
 //! representations.
 
 extern crate alloc;
@@ -7,7 +7,17 @@ use {
     alloc::collections::BinaryHeap,
     core::cmp::Reverse,
     divan::Bencher,
-    graaf::algo::dijkstra::distances,
+    graaf::{
+        algo::{
+            dijkstra::distances,
+            fixture,
+        },
+        gen::EmptyConst,
+        op::{
+            AddWeightedArc,
+            IterAllWeightedArcs,
+        },
+    },
     std::collections::HashSet,
 };
 
@@ -20,7 +30,7 @@ const HEAP: [(Reverse<usize>, usize); 1] = [(Reverse(0), 0)];
 
 #[divan::bench]
 fn vec_vec(bencher: Bencher<'_, '_>) {
-    let digraph = [vec![(1, 2)], vec![(2, 2)], Vec::new(), vec![(0, 2)]];
+    let digraph = fixture::kattis_crosscountry();
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
 
@@ -33,12 +43,10 @@ fn vec_vec(bencher: Bencher<'_, '_>) {
 
 #[divan::bench]
 fn vec_hash_set(bencher: Bencher<'_, '_>) {
-    let digraph = vec![
-        HashSet::from([(1, 2)]),
-        HashSet::from([(2, 2)]),
-        HashSet::new(),
-        HashSet::from([(0, 2)]),
-    ];
+    let digraph = fixture::kattis_crosscountry()
+        .into_iter()
+        .map(HashSet::from_iter)
+        .collect::<Vec<HashSet<(usize, usize)>>>();
 
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
@@ -52,9 +60,13 @@ fn vec_hash_set(bencher: Bencher<'_, '_>) {
 
 #[divan::bench]
 fn arr_vec(bencher: Bencher<'_, '_>) {
-    let digraph = [vec![(1, 2)], vec![(2, 2)], Vec::new(), vec![(0, 2)]];
+    let mut digraph = <[Vec<(usize, usize)>; 4]>::empty();
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
+
+    for (s, t, w) in fixture::kattis_crosscountry().iter_all_weighted_arcs() {
+        digraph.add_weighted_arc(s, t, *w)
+    }
 
     bencher.bench_local(|| {
         distances(&digraph, |acc, w| acc + w, &mut dist, &mut heap);
@@ -65,15 +77,13 @@ fn arr_vec(bencher: Bencher<'_, '_>) {
 
 #[divan::bench]
 fn arr_hash_set(bencher: Bencher<'_, '_>) {
-    let digraph = [
-        HashSet::from([(1, 2)]),
-        HashSet::from([(2, 2)]),
-        HashSet::new(),
-        HashSet::from([(0, 2)]),
-    ];
-
+    let mut digraph = <[HashSet<(usize, usize)>; 4]>::empty();
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
+
+    for (s, t, w) in fixture::kattis_crosscountry().iter_all_weighted_arcs() {
+        digraph.add_weighted_arc(s, t, *w)
+    }
 
     bencher.bench_local(|| {
         distances(&digraph, |acc, w| acc + w, &mut dist, &mut heap);