+++ benches/algo/shortestpath1.rs
@@ -1,4 +1,4 @@
-//! Benchmark [`graaf::algo::dijkstra::min_distances`] for different digraph
+//! Benchmark [`graaf::algo::dijkstra::distances`] for different digraph
 //! representations.
 
 extern crate alloc;
@@ -7,7 +7,17 @@ use {
     alloc::collections::BinaryHeap,
     core::cmp::Reverse,
     divan::Bencher,
-    graaf::algo::dijkstra::distances,
+    graaf::{
+        algo::{
+            dijkstra::distances,
+            fixture,
+        },
+        gen::EmptyConst,
+        op::{
+            AddWeightedArc,
+            IterAllWeightedArcs,
+        },
+    },
     std::collections::HashSet,
 };
 
@@ -20,13 +30,7 @@ const HEAP: [(Reverse<usize>, usize); 1] = [(Reverse(0), 0)];
 
 #[divan::bench]
 fn vec_vec(bencher: Bencher<'_, '_>) {
-    let digraph = [
-        vec![(1, 1), (2, 3), (3, 14)],
-        vec![(0, 2), (2, 4), (3, 22)],
-        vec![(0, 3), (1, 10), (3, 7)],
-        vec![(0, 13), (1, 8), (2, 2)],
-    ];
-
+    let digraph = fixture::kattis_shortestpath1();
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
 
@@ -39,12 +43,10 @@ fn vec_vec(bencher: Bencher<'_, '_>) {
 
 #[divan::bench]
 fn vec_hash_set(bencher: Bencher<'_, '_>) {
-    let digraph = vec![
-        HashSet::from([(1, 1), (2, 3), (3, 14)]),
-        HashSet::from([(0, 2), (2, 4), (3, 22)]),
-        HashSet::from([(0, 3), (1, 10), (3, 7)]),
-        HashSet::from([(0, 13), (1, 8), (2, 2)]),
-    ];
+    let digraph = fixture::kattis_shortestpath1()
+        .into_iter()
+        .map(HashSet::from_iter)
+        .collect::<Vec<HashSet<(usize, usize)>>>();
 
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
@@ -58,15 +60,13 @@ fn vec_hash_set(bencher: Bencher<'_, '_>) {
 
 #[divan::bench]
 fn arr_vec(bencher: Bencher<'_, '_>) {
-    let digraph = [
-        vec![(1, 1), (2, 3), (3, 14)],
-        vec![(0, 2), (2, 4), (3, 22)],
-        vec![(0, 3), (1, 10), (3, 7)],
-        vec![(0, 13), (1, 8), (2, 2)],
-    ];
-
+    let mut digraph = <[Vec<(usize, usize)>; 4]>::default();
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
+
+    for (s, t, w) in fixture::kattis_shortestpath1().iter_all_weighted_arcs() {
+        digraph.add_weighted_arc(s, t, *w)
+    }
 
     bencher.bench_local(|| {
         distances(&digraph, |acc, w| acc + w, &mut dist, &mut heap);
@@ -77,15 +77,13 @@ fn arr_vec(bencher: Bencher<'_, '_>) {
 
 #[divan::bench]
 fn arr_hash_set(bencher: Bencher<'_, '_>) {
-    let digraph = [
-        HashSet::from([(1, 1), (2, 3), (3, 14)]),
-        HashSet::from([(0, 2), (2, 4), (3, 22)]),
-        HashSet::from([(0, 3), (1, 10), (3, 7)]),
-        HashSet::from([(0, 13), (1, 8), (2, 2)]),
-    ];
-
+    let mut digraph = <[HashSet<(usize, usize)>; 4]>::empty();
     let mut dist = DIST;
     let mut heap = BinaryHeap::from(HEAP);
+
+    for (s, t, w) in fixture::kattis_shortestpath1().iter_all_weighted_arcs() {
+        digraph.add_weighted_arc(s, t, *w)
+    }
 
     bencher.bench_local(|| {
         distances(&digraph, |acc, w| acc + w, &mut dist, &mut heap);