+++ src/algo/mod.rs
@@ -2,9 +2,8 @@
 //!
 //! Traverse and search digraphs.
 
-mod fixture;
-
 pub mod bellman_ford_moore;
 pub mod bfs;
 pub mod dijkstra;
+pub mod fixture;
 pub mod predecessor;